создал базовый класс для карточек маршрута; вынес в него общие методы и св-ва;
